Save folder consolidation for later
