+++ src/commands/createContainerApp/createContainerApp.ts
@@ -11,7 +11,7 @@ import { ContainerAppItem } from "../../tree/ContainerAppItem";
 import type { ManagedEnvironmentItem } from "../../tree/ManagedEnvironmentItem";
 import { createActivityContext } from "../../utils/activityUtils";
 import { localize } from "../../utils/localize";
-import { containerAppEnvironmentExperience } from "../../utils/pickItem/pickContainerApp";
+import { containerAppEnvironmentExperience } from "../../utils/pickContainerApp";
 import { ImageSourceListStep } from "../imageSource/ImageSourceListStep";
 import { IngressPromptStep } from "../ingress/IngressPromptStep";
 import { ContainerAppCreateStep } from "./ContainerAppCreateStep";
