+++ src/commands/gitHub/connectToGitHub/connectToGitHub.ts
@@ -11,7 +11,7 @@ import { getGitHubAccessToken } from "../../../gitHub/getGitHubAccessToken";
 import type { ContainerAppItem } from "../../../tree/ContainerAppItem";
 import { createActivityContext } from "../../../utils/activityUtils";
 import { localize } from "../../../utils/localize";
-import { pickContainerApp } from "../../../utils/pickItem/pickContainerApp";
+import { pickContainerApp } from "../../../utils/pickContainerApp";
 import { AcrListStep } from "../../imageSource/containerRegistry/acr/AcrListStep";
 import { AcrRepositoriesListStep } from "../../imageSource/containerRegistry/acr/AcrRepositoriesListStep";
 import { DockerfileLocationInputStep } from "./DockerfileLocationInputStep";
