+++ src/commands/gitHub/openGitHubRepo.ts
@@ -8,7 +8,7 @@ import { ITreeItemPickerContext, openUrl } from "@microsoft/vscode-azext-utils";
 import type { ContainerAppsItem } from "../../tree/ContainerAppsBranchDataProvider";
 import type { ActionsTreeItem } from "../../tree/gitHub/ActionsTreeItem";
 import { localize } from "../../utils/localize";
-import { pickContainerApp } from "../../utils/pickItem/pickContainerApp";
+import { pickContainerApp } from "../../utils/pickContainerApp";
 import { getContainerAppSourceControl } from "./connectToGitHub/getContainerAppSourceControl";
 
 export async function openGitHubRepo(context: ITreeItemPickerContext, node?: ContainerAppsItem | ActionsTreeItem): Promise<void> {
