+++ src/commands/revisionDraft/discardRevisionDraft.ts
@@ -7,7 +7,7 @@ import { IActionContext } from "@microsoft/vscode-azext-utils";
 import { ext } from "../../extensionVariables";
 import { RevisionDraftItem } from "../../tree/revisionManagement/RevisionDraftItem";
 import { localize } from "../../utils/localize";
-import { pickContainerApp } from "../../utils/pickItem/pickContainerApp";
+import { pickContainerApp } from "../../utils/pickContainerApp";
 
 export async function discardRevisionDraft(context: IActionContext, node?: RevisionDraftItem): Promise<void> {
     const containerAppsItem = node ?? await pickContainerApp(context);
